requirements.txt
colorlog==6.6.0
python-dotenv==0.21.0
requests==2.28.1
setuptools==60.5.0
